requirements.txt
gradio
fastapi
uvicorn
pandas
numpy
torch
transformers
Pillow
ultralytics
scikit-learn
scikit-image
